New rb:filter/1 function to ease report filtering
Currently in the rb module the only way to filter
reports is by using the grep/1 or re/1 functions
that use Regular Expressions.

This new function allow us to specify detailed
filters that will match against our reports.

Since the reports are proplists the filters are
tuples of the form {Key, Value}. If the report
contains that tuple, it will be displayed.

Usage:

1> rb:filter([{"foo", "bar"}, {"abc", "value"}]).

2> rb:filter([{"foo", "bar", no}]).
% excludes reports containing {"foo", "bar"}

3> rb:filter([{"foo", RegExp, re}]).
% the report must contain an element with Key = "foo"
% and Value must match RegExp