compiler: keep line numbers for attributes
In the future, we might want to generate warnings
for attributes, referring to them with line numbers.

sys_pre_expand used to replace line number for attributes
with 0. Change sys_pre_expand to retain the real line
number.

v3_core used to throw away the line numbers. Change
v3_core so that it retains the line numbers in annotations.

While at it, do some tidying as suggested by tidier.